Polishing webhook for "superdesk/superdesk"

+++ bin/lxc-copy.sh
@@ -1,10 +1,20 @@
 #!/bin/sh
 set -ex
 
-from=${from:-'sd-base'}
+base=${base:-'sd-base'}
+secs=${secs-5}
+clean=${clean-1}
 
-lxc-stop -n $from || true
-lxc-stop -n $name || true
-lxc-destroy -n $name || true
-lxc-copy -n $from -N $name
-lxc-start -n $name
+if [ -n "$clean" ]; then
+    lxc-stop -n $name || true
+    lxc-destroy -n $name || true
+fi
+
+exist=$(lxc-info -n $name -sH || echo '')
+if [ -z "$exist" ]; then
+    lxc-stop -n $base && base_stoped=1 || true
+    lxc-copy -n $base -N $name
+    lxc-start -n $name
+    [ -n "$base_stoped" ] && lxc-start -n $base
+    [ -n "$secs" ] && sleep $secs
+fi
